New lib && lib fix!
