+++ Librarys/GRN.h
@@ -1,0 +1,9 @@
+#ifndef GENERATE_RANDOM_NUMBER_NUM
+#define GENERATE_RANDOM_NUMBER_NUM
+
+int getRandomNumber(int min, int max);
+
+int* getArrayOfRandomNumbers(int min, int max, int count);
+
+#include "GRN.c"
+#endif // !GENERATE_RANDOM_NUMBER_NUM
